Improve clock speed timer overhead.
Yes, this is the most stupid and obvious change in the world! The previous implementation re-read the timer interval twice every loop to cancel out the time taken by the debugger (since the whole execution thread is inside a cpu_timerMut lock, if the timer fired before the simulator ended its job the next loop would be delayed to the end of the simulator call). But the inferior limit of 1 on the sleep period makes sure that the feedback system is not destabilized; the only thing that happens is that the timer skips the first beat.